docs(lint-staged): added lint-staged
We always want to lint code before it is committed to avoid committing files that contain errors or
do not conform to our code formatting rules. We want to run this process with auto fix (`--fix`). A
problem with running this on commit is that it will potentially change files that will not be part
of the commit. To solve this we have added another cool tool
[lint-staged](https://github.com/okonet/lint-staged). This tool will only run ESLint and Prettier on
the staged files, which are the files we are committing. It will commit any files that are changed
during this process. Before it runs ESLint and Prettier it stages any unstaged and untracked files
and restores them ones the process has completed. This way these files will not be processed. Pretty
cool, right?